+++ include/mm/malloc.h
@@ -1,6 +1,7 @@
 #ifndef MALLOC_H
 #define MALLOC_H
 
+#include <stddef.h>
 #include <stdint.h>
 
 #define HEAP_SIZE   0x1000
@@ -10,15 +11,15 @@
 
 typedef struct {
   char active;
-  int idx;
-  int size;
+  size_t idx;
+  size_t size;
 } alloc_t;
 
 typedef char heap_t[HEAP_SIZE];
 typedef uint64_t bitmap_t;
 typedef alloc_t allocs_t[BITMAP_SIZE];
 
-void *kmalloc(int size);
+void *kmalloc(size_t size);
 void kfree(void *ptr);
 
 #endif
